ir: Split basic into vector+basic
