+++ devito/ir/support/__init__.py
@@ -1,3 +1,4 @@
+from devito.ir.support.vector import *  # noqa
 from devito.ir.support.basic import *  # noqa
 from devito.ir.support.space import *  # noqa
 from devito.ir.support.stencil import *  # noqa
